pyproject.toml
[tool.poetry]
name = "javacode-test-case"
version = "0.1.0"
description = ""
authors = ["Andrey Barinov <barbpro89@mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
django-wait-for-db = "^1.0.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
coverage = "^7.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
